requirements.txt
msal==1.32.3
requests==2.32.2
python-dotenv==1.0.1
